create table view for article selection
